refactor: rename openai module name
