+++ tests/models/conftest.py
@@ -4,8 +4,10 @@ from unittest.mock import AsyncMock, MagicMock, patch
 
 import aiohttp
 import pytest
+
 from readmeai.config.settings import ConfigLoader
 from readmeai.extractors.models import RepositoryContext
+from readmeai.models._openai import OpenAIHandler
 from readmeai.models.anthropic import ANTHROPIC_AVAILABLE, AnthropicHandler
 from readmeai.models.enums import GeminiModels
 from readmeai.models.gemini import GeminiHandler
